モデルの定義をファイル分割
